+++ composer.json
@@ -39,6 +39,11 @@
         "ext-sqlite3": "Allows to use an SQLite database for caching",
         "ext-dom": "Allows to use some bridges based on XPath expressions"
     },
+    "autoload-dev": {
+        "psr-4": {
+            "RssBridge\\Tests\\": "tests"
+        }
+    },
     "scripts": {
         "test": "./vendor/bin/phpunit",
         "lint": "./vendor/bin/phpcs --standard=phpcs.xml --warning-severity=0 --extensions=php -p ./",
