+++ tests/Formats/samples/expectedMrssFormat/feed.empty.xml
